Rewrite after tests

+++ Task2/gift_certificates_system/core/src/main/java/com/epam/esm/core/service/impl/GiftTagServiceImpl.java
@@ -28,12 +28,18 @@ public class GiftTagServiceImpl implements GiftTagService {
 
     @Override
     public boolean save(GiftTag giftTag) throws InvalidEntityFieldException {
+        if (giftTag == null) {
+            return false;
+        }
         validatorService.validateTag(giftTag);
         return repository.save(giftTag) > 0;
     }
 
     @Override
     public boolean delete(GiftTag giftTag) {
+        if (giftTag == null) {
+            return false;
+        }
         return repository.deleteById(findByName(giftTag.getName()).orElse(new GiftTag()).getId());
     }
 }
